tests/requirements.txt
selenium
chromedriver_installer
factory-boy
fake-factory
redislite
